// src/utils/interfaces/companyDetailInterface.ts
export interface CompanyDetails {
    industryIds: number[];
    address: string;
    city: string;
    state: string;
    country: string;
    zipCode: string;
    title?: string;
    salary?: string;
    department?: string;
    branch?: string;
    industry?: string;
    startDate?: string;
    endDate?: string;
    status?: string;
    resumeStatus?: string;
    editorContent?: string;
  }